Update README.md - Added noification for ownership transfer (#2910)
Summary:
This is a update to notify developers that the facebook-for-woocommerce plugin is now part of Meta and development is taken care by the engineers at Meta. We also have encouraged for contributions on this repo.

### Changes proposed in this Pull Request:

This PR contains mostly the changes around **README.md**.
Since mostly developers of any kind prefer looking at the **README** of a GitHub repo.
We have now changed the README emphasising the changes of the new **Owneship of the plugin** by Meta.
And the invitation for **opensource community** for their contributions.

### Screenshots:

 {F1975511490}

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/2910

Reviewed By: mshymon, vinkmeta

Differential Revision: D70323913

Pulled By: SayanPandey

fbshipit-source-id: 3fd9937d20f968132dba4fcc23ebfcabae9a3f55