Enabled rollout switch only for plugin admins (#3242)
Summary:
## Description

Primary aim of the Rollout Switches is to control the rollouts of admin functionalities, thus it should be processed only for admins and not for every website user. Also, it should be executed on hourly basis but according to current implementation it will be executed only when admin dashboard initialisation which is incorrect because if we disable rollout switch for any business then it won't get disabled until log in activity of any admin of the WordPress site. Fixed it by implementing an hearbeat.

### Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have commented my code, particularly in hard-to-understand areas.
- [x] I have confirmed that my changes do not introduce any new PHPCS warnings or errors.
- [x] I have checked plugin debug logs that my changes do not introduce any new PHP warnings or FATAL errors.
- [x] I followed general Pull Request best practices. Meta employees to follow this [wiki]([url](https://fburl.com/wiki/2cgfduwc)).
- [x] I have added tests (if necessary) and all the new and existing unit tests pass locally with my changes.
- [x] I have completed dogfooding and QA testing, or I have conducted thorough due diligence to ensure that it does not break existing functionality.
- [x] I have updated or requested update to plugin documentations (if necessary). Meta employees to follow this [wiki]([url](https://fburl.com/wiki/nhx73tgs)).

## Changelog entry

Enabled rollout switch functionality only for Plugin Admins

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/3242

Reviewed By: SayanPandey, francorisso

Differential Revision: D75409260

Pulled By: vinkmeta

fbshipit-source-id: 36bedef94674167faa49ccecfb36e6349d51c1b5