Fix to remove html encoding in fb product sets names (#3266)
Summary:
## Description

If the category name contains special characters it is by default comes in the special html format.
For example: "Clothing (v1 & v2)" becomes "Clothing (v1 &amp; v2)".
This PR is a fix to remove the html encoding before string is passed to Meta as a product set name.

### Type of change

Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)

## Changelog entry

- Fix to remove html encoding in fb product sets names

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/3266

Test Plan:
`npm run test:php`

## Screenshots
Please provide screenshots or snapshots of the system/state both before and after implementing the changes, if appropriate
### Before
<img width="1587" alt="Screenshot 2025-05-29 at 11 06 57" src="https://github.com/user-attachments/assets/f33251a7-5d6f-479c-a7e7-1cbdc12ba352" />

### After
<img width="1576" alt="Screenshot 2025-05-29 at 11 05 34" src="https://github.com/user-attachments/assets/66bf269c-f090-43fb-ad34-5a3c235bc46b" />

Reviewed By: vinkmeta, gurtejrehal

Differential Revision: D75596662

Pulled By: mshymon

fbshipit-source-id: b6e0d51af7b443269b7a679db5409eea2f61ee26