When token is invalidated because app is uninstalled we want users to be able to disconnect (#3282)
Summary:
## Description
When token is invalidated because app is uninstalled we want users to be able to disconnect from WooCommerce so that they can re-onboard. The error code from Meta for token invalidation is 190. We check if this is the error and if yes we allow disconnect to delete options from DB so that users can reconnect. We also show this error when fetching the event configs.

### Type of change
- Bug fix (non-breaking change which fixes an issue)

## Changelog entry
When token is invalidated because app is uninstalled we want users to be able to disconnect

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/3282

Test Plan: https://github.com/user-attachments/assets/e05fba58-8c30-4268-86e6-38b251fef43c

Reviewed By: woo-ardsouza

Differential Revision: D75650583

Pulled By: sharunaanandraj

fbshipit-source-id: 91694c071e64129d81f343c23caf9a2cc4fffe00