Update CAPI event handling (#3360)
Summary:
## Description

This changeset updates CAPI event handling and adds logging for Purchase events

The goal of the changes is to make tracking more robust and address deduplication issues encountered by some users.

### Type of change

- Fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have commented my code, particularly in hard-to-understand areas, if any.
- [x] I have confirmed that my changes do not introduce any new PHPCS warnings or errors.
- [x] I have checked plugin debug logs that my changes do not introduce any new PHP warnings or FATAL errors.
- [x] I followed general Pull Request best practices. Meta employees to follow this [wiki]([url](https://fburl.com/wiki/2cgfduwc)).
- [n/a] I have added tests (if necessary) and all the new and existing unit tests pass locally with my changes.
- [x] I have completed dogfooding and QA testing, or I have conducted thorough due diligence to ensure that it does not break existing functionality.
- [n/a] I have updated or requested update to plugin documentations (if necessary). Meta employees to follow this [wiki]([url](https://fburl.com/wiki/nhx73tgs)).

## Changelog entry

Update CAPI event handling & add logging for the Purchase event

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/3360

Test Plan: To test, please install this build of the plugin on a site connected to MBE and browse the site triggering the various events. Analyze the tracked events in the Events Manager, the result should be a 1:1 distribution between CAPI and pixel events being tracked.

Reviewed By: maxkozin

Differential Revision: D76722633

Pulled By: vahidkay-meta

fbshipit-source-id: fed6047c951281eea24518ed0de2e175639f7f23