Fix exclude_sale_price Value Type in Discount Syncing (#3487)
Summary:
## Description

When syncing promotions data, we currently read the exclude_ale_items field as a bool and pass that into the promotions feed as a bool value. This gets converted to an integer in the feed file for the column/field exclude_sale_priced_products.

The exclude_sale_priced_products column is actually defined to take in a 'YES' or 'NO' value representing the setting. This fixes the generation of the feed to set YES or NO.

### Type of change

Please delete options that are not relevant

- Fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have commented my code, particularly in hard-to-understand areas, if any.
- [x] I have confirmed that my changes do not introduce any new PHPCS warnings or errors.
- [x] I have checked plugin debug logs that my changes do not introduce any new PHP warnings or FATAL errors.
- [x] I followed general Pull Request best practices. Meta employees to follow this [wiki]([url](https://fburl.com/wiki/2cgfduwc)).
- [x] I have added tests (if necessary) and all the new and existing unit tests pass locally with my changes.
- [x] I have completed dogfooding and QA testing, or I have conducted thorough due diligence to ensure that it does not break existing functionality.
- [x] I have updated or requested update to plugin documentations (if necessary). Meta employees to follow this [wiki]([url](https://fburl.com/wiki/nhx73tgs)).

## Changelog entry
Fix exclude_sale_price Value Type in Discount Syncing

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/3487

Test Plan:
Ran fixed integration test
Tested syncing of discounts from WooC to Meta

Reviewed By: carterbuce

Differential Revision: D77675850

Pulled By: mradmeta

fbshipit-source-id: fe0542476319248f1f365aa40a2ff7bfb7fa54e5