Run e2e tests on marketplace release of the plugin (#3535)
Summary:
Added e2e tests on marketplace release of the plugin , in the set stable tag workflow

## Description

Updated an existing workflow - product-creation-tests.yml to become reusable , and take marketplace-version as input. If this input is passed, it pulls the marketplace version of the plugin to run the e2e tests against. Else, uses the local workspace.

Added a new job in set-stable-tag workflow , to run existing e2e tests on the marketplace version , using this reusable workflow.

### Type of change

- Add (non-breaking change which adds functionality)

## Checklist

- [✅] I have commented my code, particularly in hard-to-understand areas, if any.
- [✅] I have confirmed that my changes do not introduce any new PHPCS warnings or errors.
- [✅] I have checked plugin debug logs that my changes do not introduce any new PHP warnings or FATAL errors.
- [✅] I followed general Pull Request best practices. Meta employees to follow this [wiki]([url](https://fburl.com/wiki/2cgfduwc)).
- [✅] I have added tests (if necessary) and all the new and existing unit tests pass locally with my changes.
- [✅] I have completed dogfooding and QA testing, or I have conducted thorough due diligence to ensure that it does not break existing functionality.
- [] I have updated or requested update to plugin documentations (if necessary). Meta employees to follow this [wiki]([url](https://fburl.com/wiki/nhx73tgs)).

## Changelog entry

Added e2e tests on marketplace release of the plugin

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/3535

Test Plan:
Since this change is part of the workflow that updates the stable tag on wordpress, I commented out those parts that actually updates the stable tag , and with the remaining steps of the workflow , triggered it manually in this run - https://github.com/immadhavv/facebook-for-woocommerce/actions/runs/16626682243
It pulls in the marketplace version correctly and runs the e2e tests against it.

Negative tests when no version is passed - https://github.com/immadhavv/facebook-for-woocommerce/actions/runs/16626614538/job/47044713633

The warning seen was from the existing workflow as you can see in any old example like this - https://github.com/facebook/facebook-for-woocommerce/actions/runs/16527319140/job/46743863604?pr=3531

## Screenshots

Not applicable.

### Before

### After

Reviewed By: vahidkay-meta

Differential Revision: D79262001

Pulled By: immadhavv

fbshipit-source-id: c73ead179a4aa408638ca567135878c602df830d