Fix bug in plugin version upgrade/downgrade logic in Lifecycle (#3539)
Summary:
## Description

While dogfooding PR https://github.com/facebook/facebook-for-woocommerce/issues/3534, I cam across a bug in Lifecycle plugin version update logic:

1. Install plugin version, let's say 3.5.5
2. Install plugin version higher (uprade), let's say 3.5.6
3. Install  plugin version lower (downgrade), let's say 3.5.5
4. Install any other version of the plugin and see that plugin still thinks it's on v 3.5.6 rather than 3.5.5

The bug was that we failed to update plugin version on downgrades which leads to bugs like not running functions that supposed to be run when we start installing newer version of the plugin after this.

### Type of change

Please delete options that are not relevant

- Fix (non-breaking change which fixes an issue)

## Changelog entry

Fix bug in plugin version upgrade/downgrade logic in Lifecycle

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/3539

Test Plan: This was tested as part on a feature that required this fix PR https://github.com/facebook/facebook-for-woocommerce/issues/3534.

Reviewed By: devbodaghe, gurtejrehal

Differential Revision: D79449986

Pulled By: mshymon

fbshipit-source-id: 2422361f3191fc13c328ecfd1eb499992c8977f1